Rework reading from a framebuffer

+++ tests/query.rs
@@ -213,11 +213,11 @@ fn conditional_render_nodraw() {
                .unwrap();
     }
 
-    let data: Vec<Vec<(f32, f32, f32, f32)>> = texture.read();
+    let data: Vec<Vec<(u8, u8, u8, u8)>> = texture.read();
 
     for row in data.iter() {
         for pixel in row.iter() {
-            assert_eq!(pixel, &(0.0, 0.0, 0.0, 0.0));
+            assert_eq!(pixel, &(0, 0, 0, 0));
         }
     }
